Adding UI with few hints
